Upload state changes and progress anim (upload/save disabled)

+++ SpeechingCommon/SessionData.cs
@@ -70,7 +70,8 @@ namespace SpeechingCommon
         {
             resultsToUpload.Remove(result);
 
-            File.Delete(result.ResourceUrl);
+            if(File.Exists(result.ResourceUrl))
+                File.Delete(result.ResourceUrl);
 
             if (save) AppData.SaveCurrentData();
         }
